refactor: extract session init helpers
(cherry picked from commit 8ee357fc7688f2dbc53e832c3da2c0116a52c7f7)
